requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
distlib==0.3.9
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
dnspython==2.6.1
et-xmlfile==1.1.0
filelock==3.16.1
jwt==1.3.1
mongoengine==0.29.1
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pkg_resources==0.0.0
platformdirs==4.3.6
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
virtualenv==20.27.0
